argv and argc are how command line arguments are passed to main()
argv and argc are how command line arguments are passed to main().

argc
will be the number of strings pointed to by argv. This will be 1 plus
the number of arguments, as virtually all implementations will prepend
the name of the program to the array.

The variables are named argc
(argument count) and argv (argument vector) by convention,